College-Chatbot-main/College-Chatbot-main/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
firebase-admin==6.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
google-generativeai==0.2.0
